Moved config.py to DBConn.py
