+++ backend/db_funcs/cipherFuncs.py
@@ -1,4 +1,4 @@
-from .config import *
+from .DBConn import *
 from fastapi import APIRouter
 from routes import Cipher
